+++ et_micc2/templates/package-general-docs/{{cookiecutter.project_name}}/HISTORY.rst
@@ -5,4 +5,4 @@ History
 v{{ cookiecutter.version }} ({% now 'local' %})
 ===============================================
 
-This package was created with package `et-micc <https://github.com/etijskens/et-micc>`_,
+This package was created with package `et-micc2 <https://github.com/etijskens/et-micc2>`_,